Finish the compiler.
At this point, I think the runtime of BAM is very limited in how it
can transform the program.

We essentially need to emulate mutable references being passed across
nested contexts:
- There is no telling how much data a machine will read from a stream.
- Streams can be copied and passed down the "machine pipeline".
These two facts made it extremely difficult to devise an eager bytecode
format for BAM. Hence why the compiler only does renaming and make the
operations explicit. Which is always a good thing to have.